fix: moved exchange rate fetching functionality to crm.api.exchange_rate
(cherry picked from commit ba669d92bc70d5eb16c825c1c325e7eb51543e82)
